1.4.0 - Progress Notification

+++ src/extension.ts
@@ -7,39 +7,7 @@ const modulePath = path.join(__dirname, "..", "..", "node_modules", "ungit", "bi
 let child: ChildProcess;
 
 export class TextDocumentContentProvider implements vscode.TextDocumentContentProvider {
-    private onDidChangeEmitter = new vscode.EventEmitter<vscode.Uri>();
-    private ungitReady = false;
-
-    get onDidChange(): vscode.Event<vscode.Uri> {
-        return this.onDidChangeEmitter.event;
-    }
-
-    set ready(ready: boolean) {
-        this.ungitReady = ready;
-        this.onDidChangeEmitter.fire(previewUri);
-    }
-
     public provideTextDocumentContent(uri: vscode.Uri): string {
-        if (this.ungitReady) {
-            return this.getUngitHtml();
-        } else {
-            return this.getLoadingHtml();
-        }
-    }
-
-    private getLoadingHtml(): string {
-        const imagePath = path.join(__dirname, "..", "..", "images", "logo.png");
-        return `
-        <div style="position: fixed; height: 100%; width: 100%; margin-left: -20px; background: #252833; display: flex; justify-content: space-around; flex-direction: column; align-items: center;">
-            <image src="${imagePath}" />
-            <h1 style="color: #d8d8d8;">
-                Loading ...
-            </h1>
-        </div>
-        `;
-    }
-
-    private getUngitHtml(): string {
         const location = vscode.workspace.rootPath || "";
         const url = `http://localhost:8448/#/repository?path=${location}`;
         return `
@@ -50,33 +18,48 @@ export class TextDocumentContentProvider implements vscode.TextDocumentContentPr
     }
 }
 
-function executeCommand(provider: TextDocumentContentProvider): void {
-    vscode.commands.executeCommand("vscode.previewHtml", previewUri, vscode.ViewColumn.Two, "Ungit").then(() => {
-        return;
-    }, (reason: string) => {
-        vscode.window.showErrorMessage(reason);
-    });
-
-    child = fork(modulePath, ["--no-b", "--ungitVersionCheckOverride"], { silent: true });
-    child.stdout.on("data", (message: Buffer) => {
-        const started =
-            (message.toString().indexOf("## Ungit started ##") !== -1) ||
-            (message.toString().indexOf("Ungit server already running") !== -1);
-        if (started) {
-            provider.ready = true;
-        }
+function executeCommand(): void {
+    vscode.window.withProgress({
+        location: vscode.ProgressLocation.Notification,
+        title: "Starting Ungit",
+        cancellable: true,
+    }, (progress, token) => {
+        token.onCancellationRequested(() => {
+            if (child) {
+                child.kill();
+            }
+        });
+        return new Promise((resolve, reject) => {
+            child = fork(modulePath, ["--no-b", "--ungitVersionCheckOverride"], { silent: true });
+            child.stdout.on("data", (message: Buffer) => {
+                const started =
+                    (message.toString().includes("## Ungit started ##")) ||
+                    (message.toString().includes("Ungit server already running"));
+                if (started) {
+                    progress.report({
+                        increment: 100,
+                    });
+                    vscode.commands.executeCommand("vscode.previewHtml", previewUri, vscode.ViewColumn.Two, "Ungit").then(() => {
+                        resolve();
+                    }, (reason: string) => {
+                        vscode.window.showErrorMessage(reason);
+                        reject();
+                    });
+                }
+            });
+        });
     });
 }
 
 export function activate(context: vscode.ExtensionContext): void {
     const provider = new TextDocumentContentProvider();
     const registration = vscode.workspace.registerTextDocumentContentProvider("ungit", provider);
-    const disposable = vscode.commands.registerCommand("extension.ungit", () => {
-        executeCommand(provider);
-    });
+    const disposable = vscode.commands.registerCommand("extension.ungit", executeCommand);
     context.subscriptions.push(disposable, registration);
 }
 
 export function deactivate(): void {
-    child.kill();
+    if (child) {
+        child.kill();
+    }
 }
